Add new city

+++ database/migrations/2023_02_13_202309_create_county_cities_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('county_cities', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('county_id');
+            $table->unsignedBigInteger('county_id');
             $table->string('name')->unique();
             $table->timestamps();
